Made a lot of things unsigned for type safety.

+++ src/mod_manager.h
@@ -169,11 +169,11 @@ class mod_ui
 
         void try_add( const std::string &mod_to_add,
                       std::vector<std::string> &active_list );
-        void try_rem( int selection, std::vector<std::string> &active_list );
-        void try_shift( char direction, int &selection, std::vector<std::string> &active_list );
+        void try_rem( unsigned long selection, std::vector<std::string> &active_list );
+        void try_shift( char direction, unsigned long &selection, std::vector<std::string> &active_list );
 
-        bool can_shift_up( int selection, std::vector<std::string> active_list );
-        bool can_shift_down( int selection, std::vector<std::string> active_list );
+        bool can_shift_up( long selection, std::vector<std::string> active_list );
+        bool can_shift_down( long selection, std::vector<std::string> active_list );
 
     private:
         void set_usable_mods();
